requirements.txt
streamlit
google-generativeai
gtts
poppler-utils
streamlit-pdf-viewer
